maintain taint flag when encoding MIME on old perl
On perl 5.8.8 and below, loading utf8_heavy.pl (which happens implictly
in many places) will cause a split using a character class to not
maintain the taint flag.  Since the MIME header encoding reconstructs
the string after splitting it, we need to explicitly copy the taint flag
using a substr($s, 0, 0).