Allow 'no encoding' on EBCDIC platforms.
Commit 5ac059564287c10e630675af3b65f6e8b1103528 mistakenly reverted
afcba14d20549efcf80323ccf67c4fa36f5c16be.  This commit reinstates it.
The original commit message was:

On an EBCDIC platform, if someone specifies 'no encoding' it croaks at
compile time.  But it isn't a problem to specify 'no encoding'.  It is
only a problem to specify 'use encoding'.  So move the test to the
import function.