Precompile cp1252 code page translations
Pod::Simple now defaults to cp1252, so this translation will be used
much more frequently than before, so pre-compile it, so it doesn't have
to be generated each time.

This also causes Pod::Simple to work on EBCDIC, as there is a bug in
Encode that gets skipped when using pre-compiled translation tables.