Encode.xs: Use new core fcn if available
append_utf8_from_native_byte() is available starting in perl v5.22.  It
hides the details of this operation from the caller.  And those details
are changing in v5.23, so the existing code will no longer work.  Use
this new function when available to allow Encode to work in all
releases.