Add t/whatwg-aliases.t (check aliases against WHATWG Encoding standard)
This test checks our support for encodings and aliases against the
WHATWG Encoding Living Standard at https://encoding.spec.whatwg.org/ .

The test has many failures. I marked a few as 'TODO' but the task is
beyond my available time for now (and the aim is not to have 'TODO' but
fixes).