Add "euc-cn" => "EUC-CN" alias to Encode::MIME::Name
LWP::UserAgent uses IO::HTML for determining HTML page's encoding and
calls mime_name() method of the returned Encode::Encoding object.

This fails with Chinese webpages encoded with GB2312. IO::HTML
determined the encoding of said pages to be "euc-cn" but it's not mapped
to "EUC-CN" in Encode::MIME::Name and so the subsequent content decoding
fails.