Speed up UTF-8 validation checking on modern perls
Perl 5.26 introduced infrastructure in the core that can be used by
Encode to check UTF-8 stream validity much faster than before.

This commit replaces the current scheme for checking UTF-8 validity if
the infrastructure is availabe, by one in which normal processing
doesn't require having to decode the UTF-8 into code points.  The
copying of characters individually from the input to the output is
changed to be a single operation for each entire span of valid input at
once.

Thus in the normal case, what ends up happening is a tight loop to
check the validity, and then a memmove of the entire input to the
output, then return.

If an error is found, it copies all the valid input before the error,
then handles the character in error, then positions to the next input
position, and repeats the whole process starting from there.

Thus, this does not need to know about the intricacies of UTF-8
malformations, relying on the core to handle this.

There are currently some problems with Encode on EBCDIC platforms.  The
infrastructure is known to correctly work there, so I'm hopeful this
will solve these portability issues.

This also fixes a TODO test when Encode is run on modern perls.

Below are some benchmarking results for decode(), when called on a long
string of valid bytes which are Unicode characters consisting of 1, 2,
3, and 4 bytes.

Key:
    Ir   Instruction read
    Dr   Data read
    Dw   Data write
    COND conditional branches

The numbers represent raw counts per loop iteration.  'Inf' doesn't mean
infinity actually, but the number stored was considered too big to print
by the program that created this data.  The Ratio is the improvement of
'this' change vs. blead, and is valid even if the raw counts are shown
as 'Inf'.  The higher the ratio, the better.  The most important number
is the COND one, as each one of these interrupt the CPU instruction
pipeline.  The best result below is for 2-byte characters.  It indicates
that this patch works in less than half the time it currently takes,
i.e. the speed of this routine more than doubles with this patch on
2-byte characters, and is at least 30% faster in all cases.

decoding 1-byte
          blead     this Ratio %
       -------- -------- -------
    Ir 918975.0 319752.0   287.4
    Dr 305933.0 218330.0   140.1
    Dw 203137.0 203138.0   100.0
  COND 302803.5 227987.0   132.8

decoding 2-byte
          blead     this Ratio %
       -------- -------- -------
    Ir      Inf      Inf   413.0
    Dr      Inf 805823.0   224.1
    Dw      Inf 403139.0   248.8
  COND      Inf 902996.0   266.1

decoding 3-byte
          blead     this Ratio %
       -------- -------- -------
    Ir      Inf      Inf   246.5
    Dr      Inf      Inf   181.4
    Dw      Inf 603139.0   182.9
  COND      Inf      Inf   139.9

decoding 4-byte
          blead     this Ratio %
       -------- -------- -------
    Ir      Inf      Inf   258.7
    Dr      Inf      Inf   156.9
    Dw      Inf 803139.0   149.8
  COND      Inf      Inf   142.8