Disable -Wc++-compat when compiling with gcc
Currently when building Encode as part of core perl it produces a large
number of warnings like this:

byte_t.c:10394:24: warning: duplicate declaration of ‘utf8_viscii’ is invalid in C++ [-Wc++-compat]
10394 | static const encpage_t utf8_viscii[12] = {
      |                        ^~~~~~~~~~~
byte_t.c:322:24: note: previous declaration of ‘utf8_viscii’ was here
  322 | static const encpage_t utf8_viscii[];
      |                        ^~~~~~~~~~~

These -Wc++-compat warnings are bogus as the generated code uses define
guards to compile differently under C++ and C, the code producing the
warnings simply will not be compiled under C++ so simply silence the
warning. This patch only applies to builds under GCC, but that is
probably the bulk of our core perl builders.

Thanks to Leon Timmermans for the pragma used in this patch.

This should fix https://github.com/Perl/perl5/issues/19588 and
https://github.com/dankogai/p5-encode/issues/144