use /i matching mode consistently for file names
The first if in this sequence has a comment which says:

"VMS may have upcased filenames with DECC$ARGV_PARSE_STYLE defined"

but only the *first* clause actually uses /i to match the filename extensions.
This patch consistently uses /i for all the clauses, it also removes an
unused filename capture.